feature(refactoring): add extension of array
